+++ db2/history/reviewable_request_details.go
@@ -28,6 +28,7 @@ type ReviewableRequestDetails struct {
 	Contract              *ContractRequest          `json:"contract"`
 	AtomicSwapBidCreation *AtomicSwapBidCreation    `json:"create_atomic_swap_bid,omitempty"`
 	AtomicSwap            *AtomicSwap               `json:"create_atomic_swap,omitempty"`
+	CreatePoll            *CreatePoll               `json:"create_poll,omitempty"`
 }
 
 func (r ReviewableRequestDetails) Value() (driver.Value, error) {
@@ -156,4 +157,12 @@ type AtomicSwap struct {
 	BidID      uint64 `json:"bid_id"`
 	BaseAmount uint64 `json:"base_amount"`
 	QuoteAsset string `json:"quote_asset"`
+}
+
+type CreatePoll struct {
+	NumberOfChoices          uint32       `json:"number_of_choices"`
+	PollType                 xdr.PollType `json:"poll_type"`
+	ResultProvider           string       `json:"result_provider"`
+	VoteConfirmationRequired bool         `json:"vote_confirmation_required"`
+	Details                  map[string]interface{}
 }