+++ ingest2/generator/id_generator.go
@@ -44,3 +44,11 @@ func (g *ID) Next() int64 {
 	g.seq++
 	return (int64(g.Significant) << 32) | int64(g.seq)
 }
+
+func MakeID(significant int32, sequence uint32) int64 {
+	if significant < 0 {
+		panic(errors.New("negative significant is not supported"))
+	}
+
+	return (int64(significant) << 32) | int64(sequence)
+}